+++ chpl/chpl-service/src/main/java/gov/healthit/chpl/validation/certifiedProduct/CertifiedProductValidatorFactoryImpl.java
@@ -98,8 +98,6 @@ public class CertifiedProductValidatorFactoryImpl implements CertifiedProductVal
 		} else if(productCertificationEdition != null && productCertificationEdition.equals("2015")) {
 			return cp2015Validator;
 		} else {
-			product.getErrorMessages().add("The product is not listed as 2014 or 2015 so it cannot be validated.");
-			logger.info("Cannot find validator for certification edition '" + product.getCertificationEdition() + "'.");
 			return allowedValidator;
 		}
 		return null;